Menambahkan helper, hooks, routes, dan subject CRUD
