+++ helper/asignSubject.js
@@ -1,9 +1,9 @@
 function asignSubject(subject){
   if(subject !== null){
-    return Subject.subject_name
+    return subject.subject_name
   }else{
     return 'unassigned'
   }
 }
 
-module.exports = asiignSubject;
+module.exports = asignSubject;
