+++ models/teacher.js
@@ -1,4 +1,7 @@
 'use strict';
+// var Sequelize = require('sequelize');
+// const Op = Sequelize.Op;
+
 module.exports = (sequelize, DataTypes) => {
   var Teacher = sequelize.define('Teacher', {
     first_name: DataTypes.STRING,
@@ -8,15 +11,24 @@ module.exports = (sequelize, DataTypes) => {
       validate:{
         isEmail: {
           args: true,
-          //msg ga keluar
-          msg: 'Format bukan email'
+          msg: 'format email salah'
         },
         isEmailAvailable: function(value, msg){
-          Teacher.findAll({where:{email:value}})
+          // console.log('===========================');
+          // console.log(this.id);
+          //$notLike /$ne ===> kalo g pake Op Sequelize
+          let where = {email:value};
+          if(this.id !== null){
+            where = {email:value, id: {$ne: this.id}};
+          }
+          Teacher.findAll({where})
           .then(Results =>{
             if(Results.length != 0){
+
               msg(new Error('Email Sudah di pakai'));
               //kalo pake throw data tetep ke simpen
+            }else{
+              msg()
             }
           })
           .catch(err =>{
@@ -27,7 +39,14 @@ module.exports = (sequelize, DataTypes) => {
     },
     SubjectId: DataTypes.INTEGER
   }, {
-
+    hooks:{
+      beforeCreate: (instance, options) =>{
+        console.log('Sebelum di buat / insert');
+      },
+      afterCreate: (instance, options) =>{
+        console.log('Data Teacher Sudah di buat');
+      }
+    }
   });
   Teacher.associate = function(models) {
     // associations can be defined here
