+++ views/students.ejs
@@ -1,18 +1,11 @@
-<html>
-  <head>
-    <title>
-      Data Students
-    </title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-  </head>
-  <body>
+<% include partials/header %>
     <h2>Students Data</h2>
     <a href="/students/add">add student</a>
     <table class="table">
       <thead>
         <tr>
           <th>Id</th>
-          <th>Firt Name</th>
+          <th>First Name</th>
           <th>Last Name</th>
           <th>Email</th>
           <th>Options</th>
