+++ views/students/addStudent.ejs
@@ -14,7 +14,7 @@
     <h4>
       <a href="/students">show all data</a>
     </h4>
-    <form action="/students/addResult" method="POST">
+    <form action="/students/add" method="POST">
       <label for="first_name">First Name :</label><br>
       <input type="text" name="first_name" /><br><br>
       <label for="last_name">Last Name :</label><br>
