+++ views/subjects.ejs
@@ -1,10 +1,6 @@
-<html>
-  <head>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-  </head>
-  <body>
+<% include partials/header %>
     <h2>All Subjects Data</h2>
-    <a href="/teachers/add">Add Subject Data</a>
+    <a href="/subjects/add">Add Subject Data</a>
 
     <table class="table">
       <thead>
@@ -19,16 +15,25 @@
         <%for(let i = 0; i < Subject.length; i++) { %>
           <tr>
             <td><%= Subject[i].id %></td>
-            <td><%= Subject[i].Teachers.map(ob => `${ob.first_name} ${ob.last_name}`) %></td>
+            <td><%= Subject[i].Teachers.map(ob => ` ${ob.first_name} ${ob.last_name}`) %></td>
             <td><%= Subject[i].subject_name %></td>
             <td>
-              <a href="">Edit</a>
+              <a href="/subjects/edit/<%= Subject[i].id %>">Edit</a>
               |
-              <a href="#" onclick=" ">Delete</a>
+              <a href="#" onclick="return confirmDel(<%= Subject[i].id %>) ">Delete</a>
             </td>
           </tr>
         <% } %>
       </tbody>
     </table>
+    <script>
+      let confirmDel = function(id){
+        let confim = confirm(`Apakah kamu akan menghapus Subject Id ${id}`)
+
+        if(confirm){
+          window.location.href = `/subjects/delete/${id}`
+        }
+      }
+    </script>
   </body>
 </html>
