+++ views/teachers/addTeacher.ejs
@@ -23,7 +23,7 @@
         <%= message %>
       </div>
     <% } %>
-    <form action="/teachers/addResult" method="POST">
+    <form action="/teachers/add" method="POST">
       <label for="first_name">First Name :</label><br>
       <input type="text" name="first_name" /><br><br>
       <label for="last_name">Last Name :</label><br>
